rag_system/requirements.txt
# Dependencies for RAG System

# LLM and Vector Embeddings
openai>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.1
chromadb>=0.4.0
faiss-cpu>=1.7.4

# Document Processing
PyPDF2>=3.0.1
pandas>=2.0.0
python-docx>=0.8.11
python-dotenv>=1.0.0

# Frontend
streamlit>=1.28.0
streamlit-chat>=0.1.1

# API and Web
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# Utilities
requests>=2.31.0
numpy>=1.24.0

# Optional: For better performance and local models
sentence-transformers>=2.2.0
huggingface-hub>=0.17.0

# Development
pytest>=7.4.0
black>=23.7.0
flake8>=6.0.0
pytest-asyncio>=0.21.0
